search: move getShortInteger to utils
